requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.1.2
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.27.2
idna==3.10
injector==0.22.0
multidict==6.4.3
oauth2client==4.1.3
oauthlib==3.2.2
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyngrok==7.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.3
python-dotenv==1.0.1
python-telegram-bot[webhooks]>=22.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.19.0
asyncpg==0.30.0
